+++ src/data/records/fastq/view_traits.rs
@@ -1,14 +1,12 @@
 use crate::{
-    data::views::{Restrict, SliceRange, impl_len_for_views},
+    data::views::{Restrict, SliceRange, impl_len_for_views, impl_view_assoc_types},
     prelude::*,
 };
 
 impl_len_for_views!(FastQ, FastQView, FastQViewMut, sequence);
+impl_view_assoc_types!(FastQ, FastQView, FastQViewMut);
 
 impl DataOwned for FastQ {
-    type View<'a> = FastQView<'a>;
-    type ViewMut<'a> = FastQViewMut<'a>;
-
     #[inline]
     fn as_view(&self) -> FastQView<'_> {
         FastQView {
@@ -28,15 +26,24 @@ impl DataOwned for FastQ {
     }
 }
 
-impl DataView for FastQView<'_> {
-    type Owned = FastQ;
-
+impl<'a> DataView<'a> for FastQView<'a> {
     #[inline]
     fn to_owned_data(&self) -> FastQ {
         FastQ {
             header:   self.header.to_string(),
             sequence: self.sequence.to_owned_data(),
             quality:  self.quality.to_owned_data(),
+        }
+    }
+
+    #[inline]
+    fn reborrow_view<'b>(&'b self) -> FastQView<'b>
+    where
+        'a: 'b, {
+        FastQView {
+            header:   self.header,
+            sequence: self.sequence.reborrow_view(),
+            quality:  self.quality.reborrow_view(),
         }
     }
 }
@@ -55,14 +62,7 @@ impl Restrict for FastQView<'_> {
     }
 }
 
-impl<'b> DataViewMut<'b> for FastQViewMut<'b> {
-    type View<'a>
-        = FastQView<'a>
-    where
-        Self: 'a;
-
-    type Owned = FastQ;
-
+impl<'a> DataViewMut<'a> for FastQViewMut<'a> {
     #[inline]
     fn as_view(&self) -> FastQView<'_> {
         FastQView {
@@ -73,7 +73,7 @@ impl<'b> DataViewMut<'b> for FastQViewMut<'b> {
     }
 
     #[inline]
-    fn to_view(self) -> FastQView<'b> {
+    fn to_view(self) -> FastQView<'a> {
         FastQView {
             header:   self.header,
             sequence: self.sequence.to_view(),
@@ -87,6 +87,17 @@ impl<'b> DataViewMut<'b> for FastQViewMut<'b> {
             header:   (*self.header).clone(),
             sequence: self.sequence.to_owned_data(),
             quality:  self.quality.to_owned_data(),
+        }
+    }
+
+    #[inline]
+    fn reborrow_view_mut<'b>(&'b mut self) -> Self::ViewMut<'b>
+    where
+        'a: 'b, {
+        FastQViewMut {
+            header:   self.header,
+            sequence: self.sequence.reborrow_view_mut(),
+            quality:  self.quality.reborrow_view_mut(),
         }
     }
 }
@@ -106,8 +117,6 @@ impl Restrict for FastQViewMut<'_> {
 }
 
 impl Slice for FastQ {
-    type View<'a> = FastQView<'a>;
-
     #[inline]
     fn slice<R: SliceRange>(&self, range: R) -> FastQView<'_> {
         FastQView {
@@ -129,8 +138,6 @@ impl Slice for FastQ {
 }
 
 impl SliceMut for FastQ {
-    type ViewMut<'a> = FastQViewMut<'a>;
-
     #[inline]
     fn slice_mut<R: SliceRange>(&mut self, range: R) -> FastQViewMut<'_> {
         FastQViewMut {
@@ -151,11 +158,6 @@ impl SliceMut for FastQ {
 }
 
 impl Slice for FastQView<'_> {
-    type View<'a>
-        = FastQView<'a>
-    where
-        Self: 'a;
-
     #[inline]
     fn slice<R: SliceRange>(&self, range: R) -> FastQView<'_> {
         FastQView {
@@ -176,11 +178,6 @@ impl Slice for FastQView<'_> {
 }
 
 impl Slice for FastQViewMut<'_> {
-    type View<'a>
-        = FastQView<'a>
-    where
-        Self: 'a;
-
     #[inline]
     fn slice<R: SliceRange>(&self, range: R) -> FastQView<'_> {
         FastQView {
@@ -201,11 +198,6 @@ impl Slice for FastQViewMut<'_> {
 }
 
 impl SliceMut for FastQViewMut<'_> {
-    type ViewMut<'a>
-        = FastQViewMut<'a>
-    where
-        Self: 'a;
-
     #[inline]
     fn slice_mut<R: SliceRange>(&mut self, range: R) -> FastQViewMut<'_> {
         FastQViewMut {